+++ dst/routes/event/screeningEvent.js
@@ -25,7 +25,6 @@ const productType_1 = require("../../factory/productType");
 const TimelineFactory = require("../../factory/timeline");
 // tslint:disable-next-line:no-require-imports no-var-requires
 const subscriptions = require('../../../subscriptions.json');
-const USE_EVENT_HAS_OFFER_CATALOG = process.env.USE_EVENT_HAS_OFFER_CATALOG === '1';
 const DEFAULT_OFFERS_VALID_AFTER_START_IN_MINUTES = -20;
 var DateTimeSettingType;
 (function (DateTimeSettingType) {
@@ -1195,7 +1194,21 @@ function createEventFromBody(req) {
         });
         const superEvent = minimizeSuperEvent(screeningEventSeries);
         const eventLocation = createLocation({ id: req.project.id }, screeningRoom, maximumAttendeeCapacity);
-        return Object.assign({ project: { typeOf: req.project.typeOf, id: req.project.id }, typeOf: sdk_1.chevre.factory.eventType.ScreeningEvent, doorTime: doorTime, startDate: startDate, endDate: endDate, workPerformed: superEvent.workPerformed, location: eventLocation, superEvent: superEvent, name: superEvent.name, eventStatus: sdk_1.chevre.factory.eventStatusType.EventScheduled, offers: offers, checkInCount: undefined, attendeeCount: undefined, additionalProperty: (Array.isArray(req.body.additionalProperty))
+        return {
+            project: { typeOf: req.project.typeOf, id: req.project.id },
+            typeOf: sdk_1.chevre.factory.eventType.ScreeningEvent,
+            doorTime: doorTime,
+            startDate: startDate,
+            endDate: endDate,
+            workPerformed: superEvent.workPerformed,
+            location: eventLocation,
+            superEvent: superEvent,
+            name: superEvent.name,
+            eventStatus: sdk_1.chevre.factory.eventStatusType.EventScheduled,
+            offers: offers,
+            checkInCount: undefined,
+            attendeeCount: undefined,
+            additionalProperty: (Array.isArray(req.body.additionalProperty))
                 ? req.body.additionalProperty.filter((p) => typeof p.name === 'string' && p.name !== '')
                     .map((p) => {
                     return {
@@ -1203,15 +1216,18 @@ function createEventFromBody(req) {
                         value: String(p.value)
                     };
                 })
-                : [] }, (USE_EVENT_HAS_OFFER_CATALOG)
-            ? {
-                hasOfferCatalog: {
-                    typeOf: 'OfferCatalog',
-                    id: catalog.id,
-                    identifier: catalog.identifier
-                }
-            }
-            : undefined);
+                : []
+            // hasOfferCatalogを完全廃止(2022-09-09~)
+            // ...(USE_EVENT_HAS_OFFER_CATALOG)
+            //     ? {
+            //         hasOfferCatalog: {
+            //             typeOf: 'OfferCatalog',
+            //             id: catalog.id,
+            //             identifier: catalog.identifier
+            //         }
+            //     }
+            //     : undefined
+        };
     });
 }
 /**
@@ -1436,17 +1452,33 @@ function createMultipleEventFromBody(req) {
                     });
                     const superEvent = minimizeSuperEvent(screeningEventSeries);
                     const eventLocation = createLocation({ id: req.project.id }, screeningRoom, maximumAttendeeCapacity);
-                    attributes.push(Object.assign({ project: { typeOf: req.project.typeOf, id: req.project.id }, typeOf: sdk_1.chevre.factory.eventType.ScreeningEvent, doorTime: moment(`${formattedDate}T${data.doorTime}+09:00`, 'YYYY/MM/DDTHHmmZ')
-                            .toDate(), startDate: eventStartDate, endDate: moment(`${formattedEndDate}T${data.endTime}+09:00`, 'YYYY/MM/DDTHHmmZ')
-                            .toDate(), workPerformed: superEvent.workPerformed, location: eventLocation, superEvent: superEvent, name: superEvent.name, eventStatus: sdk_1.chevre.factory.eventStatusType.EventScheduled, offers: offers, checkInCount: undefined, attendeeCount: undefined }, (USE_EVENT_HAS_OFFER_CATALOG)
-                        ? {
-                            hasOfferCatalog: {
-                                typeOf: 'OfferCatalog',
-                                id: offerCatalog.id,
-                                identifier: offerCatalog.identifier
-                            }
-                        }
-                        : undefined));
+                    attributes.push({
+                        project: { typeOf: req.project.typeOf, id: req.project.id },
+                        typeOf: sdk_1.chevre.factory.eventType.ScreeningEvent,
+                        doorTime: moment(`${formattedDate}T${data.doorTime}+09:00`, 'YYYY/MM/DDTHHmmZ')
+                            .toDate(),
+                        startDate: eventStartDate,
+                        endDate: moment(`${formattedEndDate}T${data.endTime}+09:00`, 'YYYY/MM/DDTHHmmZ')
+                            .toDate(),
+                        workPerformed: superEvent.workPerformed,
+                        location: eventLocation,
+                        superEvent: superEvent,
+                        name: superEvent.name,
+                        eventStatus: sdk_1.chevre.factory.eventStatusType.EventScheduled,
+                        offers: offers,
+                        checkInCount: undefined,
+                        attendeeCount: undefined
+                        // hasOfferCatalogを完全廃止(2022-09-09~)
+                        // ...(USE_EVENT_HAS_OFFER_CATALOG)
+                        //     ? {
+                        //         hasOfferCatalog: {
+                        //             typeOf: 'OfferCatalog',
+                        //             id: offerCatalog.id,
+                        //             identifier: offerCatalog.identifier
+                        //         }
+                        //     }
+                        //     : undefined
+                    });
                 });
             }
         }