+++ src/routes/event/screeningEvent.ts
@@ -21,7 +21,6 @@ import * as TimelineFactory from '../../factory/timeline';
 // tslint:disable-next-line:no-require-imports no-var-requires
 const subscriptions: ISubscription[] = require('../../../subscriptions.json');
 
-const USE_EVENT_HAS_OFFER_CATALOG = process.env.USE_EVENT_HAS_OFFER_CATALOG === '1';
 const DEFAULT_OFFERS_VALID_AFTER_START_IN_MINUTES = -20;
 
 enum DateTimeSettingType {
@@ -1437,17 +1436,17 @@ async function createEventFromBody(req: Request): Promise<chevre.factory.event.s
                         value: String(p.value)
                     };
                 })
-            : [],
-        // 有無を環境変数でコントロール(2022-09-02~)
-        ...(USE_EVENT_HAS_OFFER_CATALOG)
-            ? {
-                hasOfferCatalog: {
-                    typeOf: 'OfferCatalog',
-                    id: catalog.id,
-                    identifier: catalog.identifier
-                }
-            }
-            : undefined
+            : []
+        // hasOfferCatalogを完全廃止(2022-09-09~)
+        // ...(USE_EVENT_HAS_OFFER_CATALOG)
+        //     ? {
+        //         hasOfferCatalog: {
+        //             typeOf: 'OfferCatalog',
+        //             id: catalog.id,
+        //             identifier: catalog.identifier
+        //         }
+        //     }
+        //     : undefined
     };
 }
 /**
@@ -1707,18 +1706,18 @@ async function createMultipleEventFromBody(req: Request): Promise<chevre.factory
                     eventStatus: chevre.factory.eventStatusType.EventScheduled,
                     offers: offers,
                     checkInCount: undefined,
-                    attendeeCount: undefined,
-                    // 有無を環境変数でコントロール(2022-09-02~)
-                    ...(USE_EVENT_HAS_OFFER_CATALOG)
-                        ? {
-                            hasOfferCatalog: {
-                                typeOf: 'OfferCatalog',
-                                id: offerCatalog.id,
-                                identifier: offerCatalog.identifier
-                            }
+                    attendeeCount: undefined
+                    // hasOfferCatalogを完全廃止(2022-09-09~)
+                    // ...(USE_EVENT_HAS_OFFER_CATALOG)
+                    //     ? {
+                    //         hasOfferCatalog: {
+                    //             typeOf: 'OfferCatalog',
+                    //             id: offerCatalog.id,
+                    //             identifier: offerCatalog.identifier
+                    //         }
 
-                        }
-                        : undefined
+                    //     }
+                    //     : undefined
                 });
             });
         }